HV-1444 Reorganizing code in VEManager and VEResolver
